Minor clean-up -> Added Timer_Setup.c to get Timer functions out of main.c
