+++ DPOPS_v1/DPOPS_v1.cproj
@@ -762,6 +762,9 @@
     <Compile Include="Timer_Setup.c">
       <SubType>compile</SubType>
     </Compile>
+    <Compile Include="Timer_Setup.h">
+      <SubType>compile</SubType>
+    </Compile>
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Config\" />
