+++ DPOPS_v1/Timer_Setup.h
@@ -1,0 +1,14 @@
+/*
+ * Timer_Setup.h
+ *
+ * Created: 4/1/2020 5:08:18 PM
+ *  Author: mpl055
+ */ 
+
+
+#ifndef TIMER_SETUP_H_
+#define TIMER_SETUP_H_
+
+void TIMER_0_go(void);
+
+#endif /* TIMER_SETUP_H_ */
